[Keyboard] Changes to support latest release of VIA, moved files to "idb" folder in "keymaps" in preparation of adding more boards, made changes to LAYOUT_ macros (#7798)
* make vendor folder for idb keyboards

* set default debounce to eager_pk, clean up vendor/product ID

* refactoring for VIA

* made changes to support latest release of VIA, added default keymap that's in line with the most popular configuration

* Made suggested formatting changes

* rebase to get rid of lufa files

* removed idb_60.json that's only meant for via itself

* Update keyboards/idb/idb_60/keymaps/via/rules.mk

* Update keyboards/idb/idb_60/readme.md

* Update keyboards/idb/idb_60/readme.md

* Update keyboards/idb/idb_60/readme.md

* Update keyboards/idb/idb_60/rules.mk

* Update keyboards/idb/idb_60/config.h

* Update keyboards/idb/idb_60/readme.md

* made suggested formatting changes

* fixing lufa

* Update keyboards/idb/idb_60/rules.mk

* Update keyboards/idb/idb_60/config.h

* made suggested changes
