+++ keyboards/idb/idb_60/keymaps/via/keymap.c
@@ -1,0 +1,46 @@
+#include QMK_KEYBOARD_H
+
+const uint16_t PROGMEM keymaps[][MATRIX_ROWS][MATRIX_COLS] = {
+
+  [0] = LAYOUT_default(
+    KC_GESC, KC_1,    KC_2,    KC_3,    KC_4,    KC_5,    KC_6,    KC_7,    KC_8,    KC_9,    KC_0,    KC_MINS, KC_EQL,  KC_BSPC,
+    KC_TAB,  KC_Q,    KC_W,    KC_E,    KC_R,    KC_T,    KC_Y,    KC_U,    KC_I,    KC_O,    KC_P,    KC_LBRC, KC_RBRC, KC_BSLS,
+    KC_CAPS, KC_A,    KC_S,    KC_D,    KC_F,    KC_G,    KC_H,    KC_J,    KC_K,    KC_L,    KC_SCLN, KC_QUOT,          KC_ENT,
+    KC_LSFT,          KC_Z,    KC_X,    KC_C,    KC_V,    KC_B,    KC_N,    KC_M,    KC_COMM, KC_DOT,  KC_SLSH, KC_RSFT, MO(1),
+    KC_LCTL,          KC_LALT,                                     KC_SPC,                             KC_RALT,          KC_RCTL
+  ),
+  [1] = LAYOUT_default(
+	RESET,   KC_F1,   KC_F2,   KC_F3,   KC_F4,   KC_F5,   KC_F6,   KC_F7,   KC_F8,   KC_F9,   KC_F10,  KC_F11,  KC_F12,  _______,
+	_______, _______, _______, _______, _______, _______, _______, _______, _______, _______, _______, _______, _______, _______,
+	_______, _______, _______, _______, _______, _______, _______, _______, _______, _______, _______, _______,          _______,
+	_______,          _______, _______, _______, _______, _______, _______, _______, _______, _______, _______, _______, _______,
+	_______,          _______,                                     _______,                            _______,          _______
+  ),
+  [2] = LAYOUT_default(
+	_______, _______, _______, _______, _______, _______, _______, _______, _______, _______, _______, _______, _______, _______,
+	_______, _______, _______, _______, _______, _______, _______, _______, _______, _______, _______, _______, _______, _______,
+	_______, _______, _______, _______, _______, _______, _______, _______, _______, _______, _______, _______,          _______,
+	_______,          _______, _______, _______, _______, _______, _______, _______, _______, _______, _______, _______, _______,
+	_______,          _______,                                     _______,                            _______,          _______
+  ),
+  [3] = LAYOUT_default(
+	_______, _______, _______, _______, _______, _______, _______, _______, _______, _______, _______, _______, _______, _______,
+	_______, _______, _______, _______, _______, _______, _______, _______, _______, _______, _______, _______, _______, _______,
+	_______, _______, _______, _______, _______, _______, _______, _______, _______, _______, _______, _______,          _______,
+	_______,          _______, _______, _______, _______, _______, _______, _______, _______, _______, _______, _______, _______,
+	_______,          _______,                                     _______,                            _______,          _______
+  ),
+
+};
+
+// Layer LED indicators
+
+layer_state_t layer_state_set_user(layer_state_t state) {
+    if (layer_state_cmp(state, 1)) {
+        _idb_60_esc_led_on();
+    } else {
+        _idb_60_esc_led_off();
+    }
+	
+	return state;
+}
