+++ keyboards/idb/idb_60/readme.md
@@ -8,10 +8,19 @@ Hardware Availability: [/u/omgitspngu](https://www.reddit.com/user/omgitspngu/)
 
 Make example for this keyboard (after setting up your build environment):
 
-    make idb_60:default
+    make idb/idb_60:default
 
 Or to make and flash:
 
-    make idb_60:default:dfu
+    make idb/idb_60:default:flash
+	
+To compile VIA firmware:
+
+    make idb/idb_60:via
+	
+LAYOUT_* Macros:
+
+ * *LAYOUT_default*: ANSI WKL with full Backspace
+ * *LAYOUT_all*: All split keys are available
 
 See the [build environment setup](https://docs.qmk.fm/#/getting_started_build_tools) and the [make instructions](https://docs.qmk.fm/#/getting_started_make_guide) for more information. Brand new to QMK? Start with our [Complete Newbs Guide](https://docs.qmk.fm/#/newbs).
