--- keyboards/idb_60/info.json
@@ -1,76 +1,0 @@
-{
-  "keyboard_name": "idb_60",
-  "url": "https://idb-keyboards.xyz/60",
-  "maintainer": "pngu",
-  "width": 15,
-  "height": 5,
-  "layouts": {
-    "LAYOUT": {
-      "layout": [
-		{"label":"Esc","x":0,"y":0},
-		{"label":"!\n1","x":1,"y":0},
-		{"label":"@\n2","x":2,"y":0},
-		{"label":"#\n3","x":3,"y":0},
-		{"label":"$\n4","x":4,"y":0},
-		{"label":"%\n5","x":5,"y":0},
-		{"label":"^\n6","x":6,"y":0},
-		{"label":"&\n7","x":7,"y":0},
-		{"label":"*\n8","x":8,"y":0},
-		{"label":"(\n9","x":9,"y":0},
-		{"label":")\n0","x":10,"y":0},
-		{"label":"_\n-","x":11,"y":0},
-		{"label":"+\n=","x":12,"y":0},
-		{"label":"BS","x":13,"y":0},
-		{"label":"~\n`","x":14,"y":0},
-		{"label":"Tab","x":0,"y":1,"w":1.5},
-		{"label":"Q","x":1.5,"y":1},
-		{"label":"W","x":2.5,"y":1},
-		{"label":"E","x":3.5,"y":1},
-		{"label":"R","x":4.5,"y":1},
-		{"label":"T","x":5.5,"y":1},
-		{"label":"Y","x":6.5,"y":1},
-		{"label":"U","x":7.5,"y":1},
-		{"label":"I","x":8.5,"y":1},
-		{"label":"O","x":9.5,"y":1},
-		{"label":"P","x":10.5,"y":1},
-		{"label":"{\n[","x":11.5,"y":1},
-		{"label":"}\n]","x":12.5,"y":1},
-		{"label":"|\n\\","x":13.5,"y":1,"w":1.5},
-		{"label":"Caps Lock","x":0,"y":2,"w":1.75},
-		{"label":"A","x":1.75,"y":2},
-		{"label":"S","x":2.75,"y":2},
-		{"label":"D","x":3.75,"y":2},
-		{"label":"F","x":4.75,"y":2},
-		{"label":"G","x":5.75,"y":2},
-		{"label":"H","x":6.75,"y":2},
-		{"label":"J","x":7.75,"y":2},
-		{"label":"K","x":8.75,"y":2},
-		{"label":"L","x":9.75,"y":2},
-		{"label":":\n;","x":10.75,"y":2},
-		{"label":"\"\n'","x":11.75,"y":2},
-		{"label":"Enter","x":12.75,"y":2,"w":2.25},
-		{"label":"Shift","x":0,"y":3,"w":1.25},
-		{"label":">\n<\n\n|","x":1.25,"y":3},
-		{"label":"Z","x":2.25,"y":3},
-		{"label":"X","x":3.25,"y":3},
-		{"label":"C","x":4.25,"y":3},
-		{"label":"V","x":5.25,"y":3},
-		{"label":"B","x":6.25,"y":3},
-		{"label":"N","x":7.25,"y":3},
-		{"label":"M","x":8.25,"y":3},
-		{"label":"<\n,","x":9.25,"y":3},
-		{"label":">\n.","x":10.25,"y":3},
-		{"label":"?\n/","x":11.25,"y":3},
-		{"label":"Shift","x":12.25,"y":3,"w":1.75},
-		{"label":"Fn","x":14,"y":3},
-		{"label":"Ctrl","x":0,"y":4,"w":1.5},
-		{"label":"Win","x":1.5,"y":4},
-		{"label":"Alt","x":2.5,"y":4,"w":1.5},
-		{"label":"","x":4,"y":4,"w":7},
-		{"label":"Alt","x":11,"y":4,"w":1.5},
-		{"label":"Win","x":12.5,"y":4},
-		{"label":"Ctrl","x":13.5,"y":4,"w":1.5}
-      ]
-    }
-  }
-}
